eml_validate

eml_validate processes an EML document using the XSD schema for the appropriate version of EML and determines if the document is schema-valid as defined by the XSD specification

eml_validate(eml, encoding = character(), ...)

Arguments

eml

an eml class object, file, or xml document

encoding

optional, if eml is a file path / an eml and has special characters, one can gives the encoding used by xmlParse.

...

additional arguments to eml_write, such as namespaces

Value

Whether the document is valid (logical)

Examples

 f <- system.file("xsd/test", "eml.xml", package = "EML")

 ## validate given a file name, without needing to parse first
 eml_validate(f)

#> [1] TRUE

## Validate given an "eml" object
eml <- read_eml(f)
eml_validate(eml)

#> [1] TRUE

## Can validate fragments as well, though may need the relevant namespace
dataset <- new("dataset", title = "incomplete, invalid EML")
eml_validate(dataset, namespaces = c(ds = "eml://ecoinformatics.org/dataset-2.1.1"), ns = "ds")

#> Warning: The document could not be validated.

#> [1] FALSE