get_unitList

get_unitList

get_unitList(x = read_eml(system.file("xsd/eml-2.1.1/eml-unitDictionary.xml",
  package = "EML")))

Arguments

x

a unitList object. Usually found in eml@additionalMetadata[[1]]@metadata, see details

Value

a list with two data.frames: "units", a table defining unit names, types, and conversions to SI, and "unitTypes", defining the type of unit. For instance, the unit table could define "Hertz" as a unit of unitType frequency, and the unitType define frequency as a type whose dimension is 1/time.

Details

If no unitList is provided, the function reads in the eml-unitDictionary defining all standard units and unitTypes. This provides a convenient way to look up standard units and their EML-recognized names when defining metadata, e.g. in the table passed to `set_attributes()`.

Examples

# Read in additional units defined in a EML file
f <- system.file("xsd/test/eml-datasetWithUnits.xml", package = "EML")
eml <- read_eml(f)
unitList <- get_unitList(eml@additionalMetadata[[1]]@metadata[[1]])

## Read in the definitions of standard units:
get_unitList()

#> $units
#>                                      id                                name
#> 1                         dimensionless                       dimensionless
#> 2                                second                              second
#> 3                                 meter                               meter
#> 4                              kilogram                            kilogram
#> 5                                kelvin                              kelvin
#> 6                               coulomb                             coulomb
#> 7                                ampere                              ampere
#> 8                                  mole                                mole
#> 9                               candela                             candela
#> 10                               number                              number
#> 11                           cubicMeter                          cubicMeter
#> 12                        nominalMinute                       nominalMinute
#> 13                          nominalHour                         nominalHour
#> 14                           nominalDay                          nominalDay
#> 15                          nominalWeek                         nominalWeek
#> 16                          nominalYear                         nominalYear
#> 17                      nominalLeapYear                     nominalLeapYear
#> 18                             nanogram                            nanogram
#> 19                            microgram                           microgram
#> 20                            milligram                           milligram
#> 21                            centigram                           centigram
#> 22                             decigram                            decigram
#> 23                                 gram                                gram
#> 24                             dekagram                            dekagram
#> 25                            hectogram                           hectogram
#> 26                             megagram                            megagram
#> 27                                tonne                               tonne
#> 28                                pound                               pound
#> 29                                  ton                                 ton
#> 30                              celsius                             celsius
#> 31                           fahrenheit                          fahrenheit
#> 32                            nanometer                           nanometer
#> 33                           micrometer                          micrometer
#> 34                               micron                              micron
#> 35                           millimeter                          millimeter
#> 36                           centimeter                          centimeter
#> 37                            decimeter                           decimeter
#> 38                            dekameter                           dekameter
#> 39                           hectometer                          hectometer
#> 40                            kilometer                           kilometer
#> 41                            megameter                           megameter
#> 42                             angstrom                            angstrom
#> 43                                 inch                                inch
#> 44                              Foot_US                             Foot_US
#> 45                                 foot                                foot
#> 46                      Foot_Gold_Coast                     Foot_Gold_Coast
#> 47                               fathom                              fathom
#> 48                         nauticalMile                        nauticalMile
#> 49                                 yard                                yard
#> 50                          Yard_Indian                         Yard_Indian
#> 51                          Link_Clarke                         Link_Clarke
#> 52                           Yard_Sears                          Yard_Sears
#> 53                                 mile                                mile
#> 54                           nanosecond                          nanosecond
#> 55                          microsecond                         microsecond
#> 56                          millisecond                         millisecond
#> 57                          centisecond                         centisecond
#> 58                           decisecond                          decisecond
#> 59                           dekasecond                          dekasecond
#> 60                          hectosecond                         hectosecond
#> 61                           kilosecond                          kilosecond
#> 62                           megasecond                          megasecond
#> 63                               minute                              minute
#> 64                                 hour                                hour
#> 65                            kiloliter                           kiloliter
#> 66                           microliter                          microliter
#> 67                           milliliter                          milliliter
#> 68                                liter                               liter
#> 69                               gallon                              gallon
#> 70                                quart                               quart
#> 71                               bushel                              bushel
#> 72                            cubicInch                           cubicInch
#> 73                                 pint                                pint
#> 74                               radian                              radian
#> 75                               degree                              degree
#> 76                                 grad                                grad
#> 77                            megahertz                           megahertz
#> 78                            kilohertz                           kilohertz
#> 79                                hertz                               hertz
#> 80                           millihertz                          millihertz
#> 81                               newton                              newton
#> 82                                joule                               joule
#> 83                              calorie                             calorie
#> 84                   britishThermalUnit                  britishThermalUnit
#> 85                            footPound                           footPound
#> 86                                lumen                               lumen
#> 87                                  lux                                 lux
#> 88                            becquerel                           becquerel
#> 89                                 gray                                gray
#> 90                              sievert                             sievert
#> 91                                katal                               katal
#> 92                                henry                               henry
#> 93                             megawatt                            megawatt
#> 94                             kilowatt                            kilowatt
#> 95                                 watt                                watt
#> 96                            milliwatt                           milliwatt
#> 97                             megavolt                            megavolt
#> 98                             kilovolt                            kilovolt
#> 99                                 volt                                volt
#> 100                           millivolt                           millivolt
#> 101                               farad                               farad
#> 102                                 ohm                                 ohm
#> 103                            ohmMeter                            ohmMeter
#> 104                              siemen                              siemen
#> 105                               weber                               weber
#> 106                               tesla                               tesla
#> 107                              pascal                              pascal
#> 108                          megapascal                          megapascal
#> 109                          kilopascal                          kilopascal
#> 110                          atmosphere                          atmosphere
#> 111                                 bar                                 bar
#> 112                            millibar                            millibar
#> 113             kilogramsPerSquareMeter             kilogramsPerSquareMeter
#> 114                 gramsPerSquareMeter                 gramsPerSquareMeter
#> 115            milligramsPerSquareMeter            milligramsPerSquareMeter
#> 116                 kilogramsPerHectare                 kilogramsPerHectare
#> 117                     tonnePerHectare                     tonnePerHectare
#> 118                 poundsPerSquareInch                 poundsPerSquareInch
#> 119               kilogramPerCubicMeter               kilogramPerCubicMeter
#> 120             milliGramsPerMilliLiter             milliGramsPerMilliLiter
#> 121                       gramsPerLiter                       gramsPerLiter
#> 122             milligramsPerCubicMeter             milligramsPerCubicMeter
#> 123                  microgramsPerLiter                  microgramsPerLiter
#> 124                  milligramsPerLiter                  milligramsPerLiter
#> 125             gramsPerCubicCentimeter             gramsPerCubicCentimeter
#>                           unitType                          parentSI
#> 1                    dimensionless                              <NA>
#> 2                             time                              <NA>
#> 3                           length                              <NA>
#> 4                             mass                              <NA>
#> 5                      temperature                              <NA>
#> 6                           charge                              <NA>
#> 7                          current                              <NA>
#> 8                           amount                              <NA>
#> 9                       luminosity                              <NA>
#> 10                   dimensionless                              <NA>
#> 11                          volume                              <NA>
#> 12                            time                            second
#> 13                            time                            second
#> 14                            time                            second
#> 15                            time                            second
#> 16                            time                            second
#> 17                            time                            second
#> 18                            mass                          kilogram
#> 19                            mass                          kilogram
#> 20                            mass                          kilogram
#> 21                            mass                          kilogram
#> 22                            mass                          kilogram
#> 23                            mass                          kilogram
#> 24                            mass                          kilogram
#> 25                            mass                          kilogram
#> 26                            mass                          kilogram
#> 27                            mass                          kilogram
#> 28                            <NA>                          kilogram
#> 29                            <NA>                          kilogram
#> 30                            <NA>                            kelvin
#> 31                            <NA>                            kelvin
#> 32                            <NA>                             meter
#> 33                            <NA>                             meter
#> 34                            <NA>                             meter
#> 35                            <NA>                             meter
#> 36                            <NA>                             meter
#> 37                            <NA>                             meter
#> 38                            <NA>                             meter
#> 39                            <NA>                             meter
#> 40                            <NA>                             meter
#> 41                            <NA>                             meter
#> 42                            <NA>                             meter
#> 43                            <NA>                             meter
#> 44                            <NA>                             meter
#> 45                            <NA>                             meter
#> 46                            <NA>                             meter
#> 47                            <NA>                             meter
#> 48                            <NA>                             meter
#> 49                            <NA>                             meter
#> 50                            <NA>                             meter
#> 51                            <NA>                             meter
#> 52                            <NA>                             meter
#> 53                            <NA>                             meter
#> 54                            <NA>                            second
#> 55                            <NA>                            second
#> 56                            <NA>                            second
#> 57                            <NA>                            second
#> 58                            <NA>                            second
#> 59                            <NA>                            second
#> 60                            <NA>                            second
#> 61                            <NA>                            second
#> 62                            <NA>                            second
#> 63                            <NA>                            second
#> 64                            <NA>                            second
#> 65                          volume                        cubicMeter
#> 66                          volume                        cubicMeter
#> 67                          volume                        cubicMeter
#> 68                          volume                        cubicMeter
#> 69                            <NA>                             liter
#> 70                            <NA>                             liter
#> 71                          volume                             liter
#> 72                          volume                             liter
#> 73                            <NA>                             liter
#> 74                           angle                              <NA>
#> 75                           angle                            radian
#> 76                           angle                            radian
#> 77                       frequency                             hertz
#> 78                       frequency                             hertz
#> 79                       frequency                              <NA>
#> 80                       frequency                             hertz
#> 81                           force                              <NA>
#> 82                          energy                              <NA>
#> 83                          energy                             joule
#> 84                          energy                             joule
#> 85                          energy                             joule
#> 86                      luminosity                              <NA>
#> 87                     illuminance                              <NA>
#> 88         radionucleotideActivity                              <NA>
#> 89                  specificEnergy                              <NA>
#> 90                  doseEquivalent                              <NA>
#> 91               catalyticActivity                              <NA>
#> 92                      inductance                              <NA>
#> 93                           power                              watt
#> 94                           power                              watt
#> 95                           power                              <NA>
#> 96                           power                              watt
#> 97             potentialDifference                              volt
#> 98             potentialDifference                              volt
#> 99             potentialDifference                              <NA>
#> 100            potentialDifference                              volt
#> 101                    capacitance                              <NA>
#> 102                     resistance                              <NA>
#> 103                    resistivity                              <NA>
#> 104                    conductance                              <NA>
#> 105                   magneticFlux                              <NA>
#> 106            magneticFluxDensity                              <NA>
#> 107                       pressure                              <NA>
#> 108                       pressure                            pascal
#> 109                       pressure                            pascal
#> 110                       pressure                            pascal
#> 111                       pressure                            pascal
#> 112                       pressure                            pascal
#> 113               arealMassDensity                              <NA>
#> 114               arealMassDensity           kilogramsPerSquareMeter
#> 115               arealMassDensity           kilogramsPerSquareMeter
#> 116               arealMassDensity           kilogramsPerSquareMeter
#> 117               arealMassDensity           kilogramsPerSquareMeter
#> 118               arealMassDensity           kilogramsPerSquareMeter
#> 119                    massDensity                              <NA>
#> 120                    massDensity            kilogramsPerCubicMeter
#> 121                    massDensity            kilogramsPerCubicMeter
#> 122                    massDensity            kilogramsPerCubicMeter
#> 123                    massDensity            kilogramsPerCubicMeter
#> 124                    massDensity            kilogramsPerCubicMeter
#> 125                    massDensity            kilogramsPerCubicMeter
#>           multiplierToSI constantToSI abbreviation
#> 1                   <NA>         <NA>         <NA>
#> 2                      1         <NA>          sec
#> 3                      1         <NA>            m
#> 4                      1         <NA>           kg
#> 5                      1         <NA>            K
#> 6                      1         <NA>            C
#> 7                      1         <NA>            A
#> 8                      1         <NA>          mol
#> 9                      1         <NA>           cd
#> 10                  <NA>         <NA>         <NA>
#> 11                     1         <NA>           m³
#> 12                    60         <NA>         <NA>
#> 13                  3600         <NA>         <NA>
#> 14                 86400         <NA>         <NA>
#> 15                604800         <NA>         <NA>
#> 16              31536000         <NA>         <NA>
#> 17              31622400         <NA>         <NA>
#> 18        0.000000000001         <NA>           ng
#> 19           0.000000001         <NA>           μg
#> 20              0.000001         <NA>           mg
#> 21               0.00001         <NA>           cg
#> 22                0.0001         <NA>           dg
#> 23                 0.001         <NA>            g
#> 24                  0.01         <NA>          dag
#> 25                   0.1         <NA>           hg
#> 26                  1000         <NA>           Mg
#> 27                  1000         <NA>            T
#> 28                0.4536         <NA>          lbs
#> 29              907.1999         <NA>          ton
#> 30                     1       273.18            C
#> 31                 0.556      255.402            F
#> 32           0.000000001         <NA>           nm
#> 33              0.000001         <NA>           μm
#> 34              0.000001         <NA>            μ
#> 35                 0.001         <NA>           mm
#> 36                  0.01         <NA>           cm
#> 37                   0.1         <NA>           dm
#> 38                    10         <NA>          dam
#> 39                   100         <NA>           hm
#> 40                  1000         <NA>           km
#> 41               1000000         <NA>           Mm
#> 42          0.0000000001         <NA>            Å
#> 43                0.0254         <NA>           in
#> 44                0.3048         <NA>         usft
#> 45                0.3048         <NA>           ft
#> 46             0.3047997         <NA>         gcft
#> 47                1.8288         <NA>         <NA>
#> 48                  1852         <NA>         <NA>
#> 49                0.9144         <NA>         yard
#> 50  0.914398530744440774         <NA>         <NA>
#> 51          0.2011661949         <NA>         <NA>
#> 52   0.91439841461602867         <NA>         <NA>
#> 53              1609.344         <NA>         mile
#> 54           0.000000001         <NA>         nsec
#> 55              0.000001         <NA>         μsec
#> 56                 0.001         <NA>         msec
#> 57                  0.01         <NA>         csec
#> 58                   0.1         <NA>         dsec
#> 59                    10         <NA>        dasec
#> 60                   100         <NA>         hsec
#> 61                  1000         <NA>         ksec
#> 62               1000000         <NA>         Msec
#> 63                    60         <NA>          min
#> 64                  3600         <NA>           hr
#> 65                     1         <NA>           kL
#> 66           0.000000001         <NA>           μl
#> 67              0.000001         <NA>           ml
#> 68                 0.001         <NA>            L
#> 69              3.785412         <NA>          gal
#> 70              0.946353         <NA>           qt
#> 71              0.035239         <NA>            b
#> 72        0.000016387064         <NA>          in³
#> 73              0.473176         <NA>         pint
#> 74                     1         <NA>          rad
#> 75          0.0174532924         <NA>            º
#> 76              0.015707         <NA>         grad
#> 77               1000000         <NA>          MHz
#> 78                  1000         <NA>          KHz
#> 79                     1         <NA>           Hz
#> 80              0.000001         <NA>          mHz
#> 81                     1         <NA>            N
#> 82                     1         <NA>            J
#> 83                4.1868         <NA>          cal
#> 84             1055.0559         <NA>          btu
#> 85              1.355818         <NA>         <NA>
#> 86                     1         <NA>           lm
#> 87                     1         <NA>           lx
#> 88                     1         <NA>           Bq
#> 89                     1         <NA>           Gy
#> 90                     1         <NA>           Sv
#> 91                     1         <NA>          kat
#> 92                     1         <NA>            H
#> 93               1000000         <NA>           MW
#> 94                  1000         <NA>           kW
#> 95                     1         <NA>            W
#> 96                 0.001         <NA>           mW
#> 97               1000000         <NA>           MV
#> 98                  1000         <NA>           kV
#> 99                     1         <NA>            V
#> 100                0.001         <NA>           mV
#> 101                    1         <NA>            F
#> 102                    1         <NA>            Ω
#> 103                    1         <NA>           Ωm
#> 104                    1         <NA>            S
#> 105                    1         <NA>           Wb
#> 106                    1         <NA>            T
#> 107                    1         <NA>           Pa
#> 108              1000000         <NA>          MPa
#> 109                 1000         <NA>          kPa
#> 110               101325         <NA>          atm
#> 111               100000         <NA>          bar
#> 112                  100         <NA>         mbar
#> 113                    1         <NA>        kg/m²
#> 114                0.001         <NA>         g/m²
#> 115             0.000001         <NA>        mg/m²
#> 116               0.0001         <NA>         <NA>
#> 117                  0.1         <NA>         <NA>
#> 118                17.85         <NA>      lbs/in²
#> 119                    1         <NA>         <NA>
#> 120                    1         <NA>        kg/m³
#> 121                    1         <NA>          g/l
#> 122             0.000001         <NA>        mg/m³
#> 123             0.000001         <NA>         μg/l
#> 124                0.001         <NA>         mg/l
#> 125                 1000         <NA>        g/cm³
#>                                                                                description
#> 1                                a designation asserting the absence of an associated unit
#> 2                                                                          SI unit of time
#> 3                                                                        SI unit of length
#> 4                                                                          SI unit of mass
#> 5                                                                   SI unit of temperature
#> 6                                                                        SI unit of charge
#> 7                                                            SI unit of electrical current
#> 8                                                              SI unit of substance amount
#> 9                                                                    SI unit of luminosity
#> 10                                                                                a number
#> 11                                                                             cubic meter
#> 12                                           one minute excluding leap seconds, 60 seconds
#> 13                                           one hour excluding leap seconds, 3600 seconds
#> 14                                           one day excluding leap seconds, 86400 seconds
#> 15                                          one day excluding leap seconds, 604800 seconds
#> 16                         one year excluding leap seconds and leap days, 31536000 seconds
#> 17                               one 366 day year excluding leap seconds, 31622400 seconds
#> 18                                                                       0.000000000001 kg
#> 19                                                                          0.000000001 kg
#> 20                                                                             0.000001 kg
#> 21                                                                              0.00001 kg
#> 22                                                                               0.0001 kg
#> 23                                                                                0.001 kg
#> 24                                                                                  .01 kg
#> 25                                                                                   .1 kg
#> 26                                                                                 1000 kg
#> 27                                                                     metric ton or tonne
#> 28                                             1 pound in the Avoirdupois (commerce) scale
#> 29                                                      standard US (short) ton = 2000 lbs
#> 30                                                            A common unit of temperature
#> 31                    An obsolescent unit of temperature still used in popular meteorology
#> 32                                                                       .000000001 meters
#> 33                                                                          .000001 meters
#> 34                                                                          .000001 meters
#> 35                                                                             .001 meters
#> 36                                                                              .01 meters
#> 37                                                                               .1 meters
#> 38                                                                               10 meters
#> 39                                                                              100 meters
#> 40                                                                             1000 meters
#> 41                                                                          1000000 meters
#> 42                                                                     1/10000000000 meter
#> 43                                                           An imperial measure of length
#> 44                                                                               12 inches
#> 45                                                                               12 inches
#> 46                                                                               12 inches
#> 47                                                                                  6 feet
#> 48                                                                           nautical mile
#> 49                                                                                  3 feet
#> 50  This is an ESRI unit and the multiplier comes from ESRI.  It\n    may not be accurate.
#> 51  This is an ESRI unit and the multiplier comes from ESRI.  It\n    may not be accurate.
#> 52  This is an ESRI unit and the multiplier comes from ESRI.  It\n    may not be accurate.
#> 53                                                              5280 ft or 1609.344 meters
#> 54                                                                   1/1000000 of a second
#> 55                                                                    1/100000 of a second
#> 56                                                                      1/1000 of a second
#> 57                                                                       1/100 of a second
#> 58                                                                        1/10 of a second
#> 59                                                                              10 seconds
#> 60                                                                             100 seconds
#> 61                                                                            1000 seconds
#> 62                                                                         1000000 seconds
#> 63                                                                              60 seconds
#> 64                                                                            3600 seconds
#> 65                                                                           1 cubic meter
#> 66                                                                    1/1000000 of a liter
#> 67                                                                       1/1000 of a liter
#> 68                                                                               1000 cm^3
#> 69                                                                        US liquid gallon
#> 70                                                                         US liquid quart
#> 71                                                              1 bushel = 35.23907 liters
#> 72                                                                              cubic inch
#> 73                                                                          US liquid pint
#> 74                                                    2 pi radians comprise a unit circle.
#> 75                                                      360 degrees comprise a unit circle
#> 76                                      a plane angle equivalent to 1/400 of a full circle
#> 77                                                                               megahertz
#> 78                                                                               kilohertz
#> 79                                                                                   hertz
#> 80                                                                              millihertz
#> 81                                                                                  newton
#> 82                                                                             joule = N*m
#> 83                                                                        1 cal = 4.1868 J
#> 84                                                                     1 btu = 1055.0559 J
#> 85                                                                   1 ft-lbs = 1.355818 J
#> 86                                                                                   lumen
#> 87                                                                                     lux
#> 88                                                                               becquerel
#> 89                                                                                    gray
#> 90                                                                                 sievert
#> 91                                                                                   katal
#> 92                                                                                   henry
#> 93                                                                                megawatt
#> 94                                                                                kilowatt
#> 95                                                                              watt = J/s
#> 96                                                                               milliwatt
#> 97                                                                                megavolt
#> 98                                                                                kilovolt
#> 99                                                                                    volt
#> 100                                                                              millivolt
#> 101                                                                                  farad
#> 102                                                                                    ohm
#> 103                                                                             ohm meters
#> 104                                                                                 siemen
#> 105                                                                                  weber
#> 106                                                                                  tesla
#> 107                                                                                 pascal
#> 108                                                                             megapascal
#> 109                                                                             kilopascal
#> 110                                                         1 atmosphere = 101325 pascals 
#> 111                                                                 1 bar = 100000 pascals
#> 112                                                                               millibar
#> 113                                                             kilograms per square meter
#> 114                                                                 grams per square meter
#> 115                                                            milligrams Per Square Meter
#> 116                                                                  kilograms per hectare
#> 117                                                        metric ton or tonne per hectare
#> 118                                                                        lbs/square inch
#> 119                                                               kilogram per cubic meter
#> 120                                                              milligrams per milliliter
#> 121                                                                        grams per liter
#> 122                                                             milligrams Per Cubic Meter
#> 123                                                                     micrograms / liter
#> 124                                                                     milligrams / liter
#> 125                                                             grams per cubic centimeter
#>  [ reached getOption("max.print") -- omitted 70 rows ]
#> 
#> $unitTypes
#>                                 id                           name     dimension
#> 1                           length                         length        length
#> 2                             time                           time          time
#> 3                             mass                           mass          mass
#> 4                          current                        current       current
#> 5                      temperature                    temperature   temperature
#> 6                           amount                         amount        amount
#> 7                       luminosity                     luminosity    luminosity
#> 8                    dimensionless                  dimensionless dimensionless
#> 9                            angle                          angle         angle
#> 10                    acceleration                   acceleration        length
#> 11                    acceleration                   acceleration          time
#> 12                          charge                         charge       current
#> 13                          charge                         charge          time
#> 14           magneticFieldStrength          magneticFieldStrength       current
#> 15           magneticFieldStrength          magneticFieldStrength        length
#> 16                  currentDensity                 currentDensity       current
#> 17                  currentDensity                 currentDensity        length
#> 18                            area                           area        length
#> 19                lengthReciprocal               lengthReciprocal        length
#> 20                       frequency                      frequency          time
#> 21                          volume                         volume        length
#> 22                  volumetricArea                 volumetricArea        length
#> 23                  volumetricArea                 volumetricArea        length
#> 24                           speed                          speed        length
#> 25                           speed                          speed          time
#> 26                     massDensity                    massDensity          mass
#> 27                     massDensity                    massDensity        length
#> 28                     massPerMass                    massPerMass          mass
#> 29                     massPerMass                    massPerMass          mass
#> 30                 volumePerVolume                volumePerVolume        length
#> 31                 volumePerVolume                volumePerVolume          mass
#> 32                  volumetricRate                 volumetricRate        length
#> 33                  volumetricRate                 volumetricRate          time
#> 34       volumetricMassDensityRate      volumetricMassDensityRate          mass
#> 35       volumetricMassDensityRate      volumetricMassDensityRate        length
#> 36       volumetricMassDensityRate      volumetricMassDensityRate          time
#> 37            arealMassDensityRate           arealMassDensityRate          mass
#> 38            arealMassDensityRate           arealMassDensityRate        length
#> 39            arealMassDensityRate           arealMassDensityRate          time
#> 40                  specificVolume                 specificVolume          mass
#> 41                  specificVolume                 specificVolume        length
#> 42  amountOfSubstanceConcentration amountOfSubstanceConcentration        amount
#> 43  amountOfSubstanceConcentration amountOfSubstanceConcentration        length
#> 44         amountOfSubstanceWeight        amountOfSubstanceWeight        amount
#> 45         amountOfSubstanceWeight        amountOfSubstanceWeight          mass
#> 46     amountOfSubstanceWeightFlux    amountOfSubstanceWeightFlux        amount
#> 47     amountOfSubstanceWeightFlux    amountOfSubstanceWeightFlux          mass
#> 48     amountOfSubstanceWeightFlux    amountOfSubstanceWeightFlux          time
#> 49                        massFlux                       massFlux          mass
#> 50                        massFlux                       massFlux          time
#> 51                       luminance                      luminance    luminosity
#> 52                       luminance                      luminance        length
#> 53               volumetricDensity              volumetricDensity dimensionless
#> 54               volumetricDensity              volumetricDensity        length
#> 55                    arealDensity                   arealDensity dimensionless
#> 56                    arealDensity                   arealDensity        length
#> 57                arealMassDensity               arealMassDensity          mass
#> 58                arealMassDensity               arealMassDensity        length
#> 59                    specificArea                   specificArea          mass
#> 60                    specificArea                   specificArea        length
#> 61                           force                          force          mass
#> 62                           force                          force        length
#> 63                           force                          force          time
#> 64                          energy                         energy          mass
#> 65                          energy                         energy        length
#> 66                          energy                         energy          time
#> 67                           power                          power          mass
#> 68                           power                          power        length
#> 69                           power                          power          time
#> 70             potentialDifference            potentialDifference          mass
#> 71             potentialDifference            potentialDifference        length
#> 72             potentialDifference            potentialDifference          time
#> 73             potentialDifference            potentialDifference       current
#> 74                     capacitance                    capacitance          mass
#> 75                     capacitance                    capacitance        length
#> 76                     capacitance                    capacitance          time
#> 77                     capacitance                    capacitance       current
#> 78                      resistance                     resistance          mass
#> 79                      resistance                     resistance        length
#> 80                      resistance                     resistance          time
#> 81                      resistance                     resistance       current
#> 82                     resistivity                    resistivity          mass
#> 83                     resistivity                    resistivity        length
#> 84                     resistivity                    resistivity          time
#> 85                     resistivity                    resistivity       current
#> 86                     conductance                    conductance          mass
#> 87                     conductance                    conductance        length
#> 88                     conductance                    conductance          time
#> 89                     conductance                    conductance       current
#> 90                    magneticFlux                   magneticFlux          mass
#> 91                    magneticFlux                   magneticFlux        length
#> 92                    magneticFlux                   magneticFlux          time
#> 93                    magneticFlux                   magneticFlux       current
#> 94             magneticFluxDensity            magneticFluxDensity          mass
#> 95             magneticFluxDensity            magneticFluxDensity          time
#> 96             magneticFluxDensity            magneticFluxDensity       current
#> 97                      inductance                     inductance          mass
#> 98                      inductance                     inductance        length
#> 99                      inductance                     inductance          time
#> 100                     inductance                     inductance       current
#> 101                    illuminance                    illuminance    luminosity
#> 102                    illuminance                    illuminance        length
#> 103        radionucleotideActivity        radionucleotideActivity          time
#> 104                 specificEnergy                 specificEnergy          time
#> 105                 specificEnergy                 specificEnergy        length
#> 106                 doseEquivalent                 doseEquivalent          time
#> 107                 doseEquivalent                 doseEquivalent        length
#> 108              catalyticActivity              catalyticActivity          time
#> 109              catalyticActivity              catalyticActivity        amount
#> 110                       pressure                       pressure          mass
#> 111                       pressure                       pressure          time
#> 112                       pressure                       pressure        length
#> 113                 transmissivity                 transmissivity        length
#> 114                 transmissivity                 transmissivity          time
#> 115             massSpecificLength             massSpecificLength        length
#> 116             massSpecificLength             massSpecificLength          mass
#> 117              massSpecificCount              massSpecificCount dimensionless
#> 118              massSpecificCount              massSpecificCount          mass
#>     power
#> 1    <NA>
#> 2    <NA>
#> 3    <NA>
#> 4    <NA>
#> 5    <NA>
#> 6    <NA>
#> 7    <NA>
#> 8    <NA>
#> 9    <NA>
#> 10   <NA>
#> 11     -2
#> 12   <NA>
#> 13      1
#> 14   <NA>
#> 15     -1
#> 16   <NA>
#> 17     -2
#> 18      2
#> 19     -1
#> 20     -1
#> 21      3
#> 22      3
#> 23     -2
#> 24   <NA>
#> 25     -1
#> 26   <NA>
#> 27     -3
#> 28   <NA>
#> 29     -1
#> 30      3
#> 31     -3
#> 32      3
#> 33     -1
#> 34   <NA>
#> 35     -3
#> 36     -1
#> 37   <NA>
#> 38     -2
#> 39     -1
#> 40     -1
#> 41      3
#> 42   <NA>
#> 43     -3
#> 44   <NA>
#> 45     -1
#> 46   <NA>
#> 47     -1
#> 48     -1
#> 49   <NA>
#> 50     -1
#> 51   <NA>
#> 52     -2
#> 53   <NA>
#> 54     -3
#> 55   <NA>
#> 56     -2
#> 57   <NA>
#> 58     -2
#> 59     -1
#> 60      2
#> 61   <NA>
#> 62   <NA>
#> 63     -2
#> 64   <NA>
#> 65      2
#> 66     -2
#> 67   <NA>
#> 68      2
#> 69     -3
#> 70   <NA>
#> 71      2
#> 72     -3
#> 73     -1
#> 74     -1
#> 75     -2
#> 76      4
#> 77      2
#> 78   <NA>
#> 79      2
#> 80     -3
#> 81     -2
#> 82   <NA>
#> 83      3
#> 84     -3
#> 85     -2
#> 86     -1
#> 87     -2
#> 88      3
#> 89      2
#> 90   <NA>
#> 91      2
#> 92     -2
#> 93     -1
#> 94   <NA>
#> 95     -2
#> 96     -1
#> 97   <NA>
#> 98      2
#> 99     -2
#> 100    -2
#> 101  <NA>
#> 102    -2
#> 103    -1
#> 104    -2
#> 105     2
#> 106    -2
#> 107     2
#> 108    -1
#> 109  <NA>
#> 110  <NA>
#> 111    -2
#> 112     1
#> 113     2
#> 114    -1
#> 115  <NA>
#> 116    -1
#> 117  <NA>
#> 118    -1
#>