set_attributes

set_attributes

set_attributes(attributes, factors = NULL, col_classes = NULL)

Arguments

attributes

a joined table of all attribute metadata (such as returned by get_attributes, see details)

factors

a table with factor code-definition pairs; see details

col_classes

optional, list of R column classes ('ordered', 'numeric', 'factor', 'Date', or 'character', case sensitive) will let the function infer missing 'domain' and 'measurementScale' values for attributes column. Should be in same order as attributeNames in the attributes table, or be a named list with names corresponding to attributeNames in the attributes table.

Value

an eml "attributeList" object

Details

The attributes data frame must use only the recognized column headers shown here. The attributes data frame must contain columns for required metadata. These are:

For all data: - attributeName (required, free text field) - attributeDefinition (required, free text field) - measurementScale (required, "nominal", "ordinal", "ratio", "interval", or "dateTime", case sensitive) but it can be inferred from col_classes. - domain (required, "numericDomain", "textDomain", "enumeratedDomain", or "dateTimeDomain", case sensitive) but it can be inferred from col_classes.

For numeric (ratio or interval) data: - unit (required)

For character (textDomain) data: - definition (required)

For dateTime data: - formatString (required)

For factor data: