eml_get

eml_get

eml_get(x, element = NULL, ...)

# S4 method for attributeList
eml_get(x, eml = x)

  # S4 method for unitList
eml_get(x = read_eml(system.file("xsd/eml-2.1.1/eml-unitDictionary.xml",
  package = "EML")))

Arguments

x

an EML object or child/descendant object

element

name of the element to be extracted. If multiple occurrences are found, will extract all

...

additional arguments for specific get_* methods, see details

eml

the eml file from which to look up references (optional)

Details

Some elements have custom get_<class> functions which return the requested metadata in a more natural format, such as get_attributeList, which returns data.frames containing the desired metadata. If such a function exists for the element requested, that function is called instead. Additional arguments given to `eml_get` will be passed to that function via ..., for example, compare get_attributes to the example shown here.

Methods (by class)

attributeList: get method for attributeList

unitList: get method for UnitList

Examples

f <- system.file("xsd/test", "eml-datasetWithUnits.xml", package = "EML")
eml <- read_eml(f)
eml_get(eml, "physical")

#> An object of class "ListOfphysical"
#> [[1]]
#> <physical>
#>   <objectName>CDR LTER-patterns among communities.txt</objectName>
#>   <size unit="bytes">1245</size>
#>   <characterEncoding>ASCII</characterEncoding>
#>   <dataFormat>
#>     <textFormat>
#>       <numHeaderLines>1</numHeaderLines>
#>       <attributeOrientation>column</attributeOrientation>
#>       <simpleDelimited>
#>         <fieldDelimiter>\t</fieldDelimiter>
#>       </simpleDelimited>
#>     </textFormat>
#>   </dataFormat>
#>   <distribution>
#>     <online>
#>       <url>http://metacat.nceas.ucsb.edu/knb/servlet/metacat?action=read&amp;docid=knb.46.1</url>
#>     </online>
#>   </distribution>
#> </physical> 
#> 

## slower examples
eml_get(eml, "attributeList")

#> $attributes
#>    attributeName         domain precision    minimum    maximum
#> 1            fld     textDomain        NA       <NA>       <NA>
#> 2           year dateTimeDomain      1.00       1944       <NA>
#> 3             sr  numericDomain      0.50          0       <NA>
#> 4         pctcov  numericDomain      0.10          0        1e2
#> 5        avesr91  numericDomain      0.10          0       <NA>
#> 6        avesr92  numericDomain      0.10          0       <NA>
#> 7        avesr93  numericDomain      0.10          0       <NA>
#> 8        avesr94  numericDomain      0.10          0       <NA>
#> 9        avesr95  numericDomain      0.10          0       <NA>
#> 10       avesr96  numericDomain      0.10          0       <NA>
#> 11        MeanSR  numericDomain      0.10          0       <NA>
#> 12       biomass  numericDomain      0.01          0       <NA>
#> 13         sppm2  numericDomain      0.01          0       <NA>
#> 14          time dateTimeDomain      0.10 15:00:00.0 19:00:00.0
#>                     unit numberType formatString            definition pattern
#> 1                   <NA>       <NA>         <NA> Valid names of fields    <NA>
#> 2                   <NA>       <NA>         YYYY                  <NA>    <NA>
#> 3                 number       real         <NA>                  <NA>    <NA>
#> 4          dimensionless       real         <NA>                  <NA>    <NA>
#> 5                 second       real         <NA>                  <NA>    <NA>
#> 6     cubicFeetPerSecond       real         <NA>                  <NA>    <NA>
#> 7             cubicMeter       real         <NA>                  <NA>    <NA>
#> 8                 radian       real         <NA>                  <NA>    <NA>
#> 9                  meter       real         <NA>                  <NA>    <NA>
#> 10                  inch       real         <NA>                  <NA>    <NA>
#> 11              nanogram       real         <NA>                  <NA>    <NA>
#> 12   gramsPerSquareMeter       real         <NA>                  <NA>    <NA>
#> 13 speciesPerSquareMeter       real         <NA>                  <NA>    <NA>
#> 14                  <NA>       <NA>   hh:mm:ss.s                  <NA>    <NA>
#>    source                    attributeLabel storageType missingValueCode
#> 1    <NA>                             Field      string             <NA>
#> 2    <NA>                              year       gYear             <NA>
#> 3    <NA>                  Species Richness       float             <NA>
#> 4    <NA>                     percent cover       float             <NA>
#> 5    <NA> Average Species Richness for 1991       float             <NA>
#> 6    <NA> Average Species Richness for 1992       float             <NA>
#> 7    <NA> Average Species Richness for 1993       float             <NA>
#> 8    <NA> Average Species Richness for 1994       float             <NA>
#> 9    <NA> Average Species Richness for 1995       float             <NA>
#> 10   <NA> Average Species Richness for 1996       float             <NA>
#> 11   <NA>             mean species richness       float             <NA>
#> 12   <NA>                           Biomass       float             <NA>
#> 13   <NA>          Species Per Square Meter       float             <NA>
#> 14   <NA>                              Time        time             <NA>
#>    missingValueCodeExplanation measurementScale
#> 1                         <NA>          nominal
#> 2                         <NA>         dateTime
#> 3                         <NA>         interval
#> 4                         <NA>            ratio
#> 5                         <NA>            ratio
#> 6                         <NA>            ratio
#> 7                         <NA>            ratio
#> 8                         <NA>            ratio
#> 9                         <NA>            ratio
#> 10                        <NA>            ratio
#> 11                        <NA>            ratio
#> 12                        <NA>            ratio
#> 13                        <NA>            ratio
#> 14                        <NA>         dateTime
#>                                              attributeDefinition
#> 1                   Field where the data was collected\n        
#> 2                      The year the data was collected\n        
#> 3                                       Species richness for CDR
#> 4                The percent ground cover on the field\n        
#> 5   The average species richness for the field in 1991\n        
#> 6   The average species richness for the field in 1992\n        
#> 7   The average species richness for the field in 1993\n        
#> 8   The average species richness for the field in 1994\n        
#> 9   The average species richness for the field in 1995\n        
#> 10  The average species richness for the field in 1996\n        
#> 11         the mean species richness from 1991 to 1996\n        
#> 12            The total biomass measured in this field\n        
#> 13                 Calculated species per square meter\n        
#> 14 The time of day for this observation, 24 hour clock\n        
#> 
#> $factors
#> NULL
#> 

## The first argument need not be an "eml" class, it could be a child element; e.g.
eml_get(eml@dataset@dataTable, "physical")

#> An object of class "ListOfphysical"
#> [[1]]
#> <physical>
#>   <objectName>CDR LTER-patterns among communities.txt</objectName>
#>   <size unit="bytes">1245</size>
#>   <characterEncoding>ASCII</characterEncoding>
#>   <dataFormat>
#>     <textFormat>
#>       <numHeaderLines>1</numHeaderLines>
#>       <attributeOrientation>column</attributeOrientation>
#>       <simpleDelimited>
#>         <fieldDelimiter>\t</fieldDelimiter>
#>       </simpleDelimited>
#>     </textFormat>
#>   </dataFormat>
#>   <distribution>
#>     <online>
#>       <url>http://metacat.nceas.ucsb.edu/knb/servlet/metacat?action=read&amp;docid=knb.46.1</url>
#>     </online>
#>   </distribution>
#> </physical> 
#>