get_attributes

get_attributes

get_attributes(x, eml = x)

Arguments

x

an "attributeList" element from EML

eml

The full eml document, needed only if <references> outside of attributes must be resolved.

Value

a data frame whose rows are the attributes (names of each column in the data file) and whose columns describe metadata about those attributes. By default separate tables are given for each type

Details

EML metadata can use "references" elements which allow one attribute to use metadata declared elsewhere in the document. This function will automatically resolve these references and thus infer the correct metadata.

Examples

 # runs > 5s

eml <- read_eml(system.file("xsd/test/eml-datasetWithAttributelevelMethods.xml", package = "EML"))
get_attributes( eml@dataset@dataTable[[1]]@attributeList )

#> $attributes
#>    attributeName         domain precision    minimum    maximum          unit
#> 1            fld     textDomain        NA       <NA>       <NA>          <NA>
#> 2           year dateTimeDomain       1.0       1944       <NA>          <NA>
#> 3             sr  numericDomain       0.5          0       <NA> dimensionless
#> 4         pctcov  numericDomain       0.1          0        100 dimensionless
#> 5        avesr91  numericDomain       0.1          0       <NA> dimensionless
#> 6        avesr92  numericDomain       0.1          0       <NA> dimensionless
#> 7        avesr93  numericDomain       0.1          0       <NA> dimensionless
#> 8        avesr94  numericDomain       0.1          0       <NA> dimensionless
#> 9        avesr95  numericDomain       0.1          0       <NA> dimensionless
#> 10       avesr96  numericDomain       0.1          0       <NA> dimensionless
#> 11        MeanSR  numericDomain       0.1          0       <NA> dimensionless
#> 12          time dateTimeDomain       0.1 15:00:00.0 19:00:00.0          <NA>
#>    numberType formatString            definition pattern source
#> 1        <NA>         <NA> Valid names of fields    <NA>   <NA>
#> 2        <NA>         YYYY                  <NA>    <NA>   <NA>
#> 3        real         <NA>                  <NA>    <NA>   <NA>
#> 4        real         <NA>                  <NA>    <NA>   <NA>
#> 5        real         <NA>                  <NA>    <NA>   <NA>
#> 6        real         <NA>                  <NA>    <NA>   <NA>
#> 7        real         <NA>                  <NA>    <NA>   <NA>
#> 8        real         <NA>                  <NA>    <NA>   <NA>
#> 9        real         <NA>                  <NA>    <NA>   <NA>
#> 10       real         <NA>                  <NA>    <NA>   <NA>
#> 11       real         <NA>                  <NA>    <NA>   <NA>
#> 12       <NA>   hh:mm:ss.s                  <NA>    <NA>   <NA>
#>                       attributeLabel storageType missingValueCode
#> 1                              Field      string             <NA>
#> 2                               year       gYear             <NA>
#> 3                   Species Richness       float             <NA>
#> 4                      percent cover       float             <NA>
#> 5  Average Species Richness for 1991       float             <NA>
#> 6  Average Species Richness for 1992       float             <NA>
#> 7  Average Species Richness for 1993       float             <NA>
#> 8  Average Species Richness for 1994       float             <NA>
#> 9  Average Species Richness for 1995       float             <NA>
#> 10 Average Species Richness for 1996       float             <NA>
#> 11             mean species richness       float             <NA>
#> 12                              Time        time             <NA>
#>    missingValueCodeExplanation measurementScale
#> 1                         <NA>          nominal
#> 2                         <NA>         dateTime
#> 3                         <NA>         interval
#> 4                         <NA>            ratio
#> 5                         <NA>            ratio
#> 6                         <NA>            ratio
#> 7                         <NA>            ratio
#> 8                         <NA>            ratio
#> 9                         <NA>            ratio
#> 10                        <NA>            ratio
#> 11                        <NA>            ratio
#> 12                        <NA>         dateTime
#>                                              attributeDefinition
#> 1                   Field where the data was collected\n        
#> 2                      The year the data was collected\n        
#> 3                                       Species richness for CDR
#> 4                The percent ground cover on the field\n        
#> 5   The average species richness for the field in 1991\n        
#> 6   The average species richness for the field in 1992\n        
#> 7   The average species richness for the field in 1993\n        
#> 8   The average species richness for the field in 1994\n        
#> 9   The average species richness for the field in 1995\n        
#> 10  The average species richness for the field in 1996\n        
#> 11         the mean species richness from 1991 to 1996\n        
#> 12 The time of day for this observation, 24 hour clock\n        
#> 
#> $factors
#> NULL
#>