set_TextType

For any EML element of class TextType, this function can be used to generate the appropriate EML from a markdown-formatted file.

set_TextType(file = NULL, text = NULL)

Arguments

file

path to a file providing formatted input text, see details.

text

a plain text character string which will be used directly as the content of the node if no file is given

Value

a TextType object that can be coerced into any element inheriting from TextType, see examples

Details

If the `rmarkdown` package is installed, then the input file can be a Microsoft Word (.docx) file, a markdown file, or other file recognized by Pandoc (see http://pandoc.org), which will automate the conversion to a docbook. Otherwise, the input file should already be in docbook format (with .xml or .dbk extension). Note that pandoc comes pre-installed in RStudio and is required for the rmarkdown package.

Examples

## using a simple character string
a <- set_TextType(text = "This is the abstract")
as(a, "abstract")

#> <abstract>This is the abstract</abstract> 

## Using an external markdown file
f <- system.file("examples/hf205-abstract.md", package = "EML")
a <- set_TextType(f)
as(a, "abstract")

#> <abstract>
#>   <para>
#>   The primary goal of this project is to determine experimentally the
#>   amount of lead time required to prevent a state change. To achieve
#>   this goal, we will (1) experimentally induce state changes in a
#>   natural aquatic ecosystem - the Sarracenia microecosystem; (2) use
#>   proteomic analysis to identify potential indicators of states and
#>   state changes; and (3) test whether we can forestall state changes by
#>   experimentally intervening in the system. This work uses state-of-the
#>   art molecular tools to identify early warning indicators in the field
#>   of aerobic to anaerobic state changes driven by nutrient enrichment in
#>   an aquatic ecosystem. The study tests two general hypotheses: (1)
#>   proteomic biomarkers can function as reliable indicators of impending
#>   state changes and may give early warning before increasing variances
#>   and statistical flickering of monitored variables; and (2) well-timed
#>   intervention based on proteomic biomarkers can avert future state
#>   changes in ecological systems.
#> </para>
#> </abstract> 

## Can also import from methods written in a .docx MS Word file.
f <- system.file("examples/hf205-abstract.docx", package = "EML")
a <- set_TextType(f)
as(a, "abstract")

#> <abstract>
#>   <para>
#>   The primary goal of this project is to determine experimentally the
#>   amount of lead time required to prevent a state change. To achieve
#>   this goal, we will (1) experimentally induce state changes in a
#>   natural aquatic ecosystem - the Sarracenia microecosystem; (2) use
#>   proteomic analysis to identify potential indicators of states and
#>   state changes; and (3) test whether we can forestall state changes by
#>   experimentally intervening in the system. This work uses state-of-the
#>   art molecular tools to identify early warning indicators in the field
#>   of aerobic to anaerobic state changes driven by nutrient enrichment in
#>   an aquatic ecosystem. The study tests two general hypotheses: (1)
#>   proteomic biomarkers can function as reliable indicators of impending
#>   state changes and may give early warning before increasing variances
#>   and statistical flickering of monitored variables; and (2) well-timed
#>   intervention based on proteomic biomarkers can avert future state
#>   changes in ecological systems.
#> </para>
#> </abstract> 

## Documents with title headings use `section` instead of `para` notation
f <- system.file("examples/hf205-methods.docx", package = "EML")
d <- set_TextType(f)
as(d, "description")

#> <description>
#>   <section>
#>     <title>General Protocols</title>
#>     <para>
#>     Field methods. All experiments will be carried out in the greenhouse
#>     at Harvard Forest. We have developed an instrumentation system that
#>     allows us to collect continuous dissolved [O2] measurements:
#>     dedicated micro-probes (DO-166MT; Lazar Research Laboratories:
#>     http://www.shelfscientific.com/) connected to multiplexers and data
#>     loggers (AM16/32B multiplexer, CR-1000 datalogger and control system
#>     [Campbell Scientific: http://www.cambellsci.com]). The initial
#>     ecosystem composition in all experimental plants will be
#>     standardized by seeding each pitcher with a 10-ml inoculum of liquid
#>     collected from pitchers growing at Tom Swamp. In all experiments,
#>     prey will be supplied to pitchers as standardized aliquots of dried
#>     and finely ground bald-faced hornets (Dolichovespula maculata;
#>     Hymenoptera: Vespidae), which we collect in quantity throughout New
#>     England. Both hornets and ants (the latter are the dominant prey of
#>     S. purpurea) are hymenoptera, and have nearly identical C:N ratios
#>     (hornets: 3.97; common bog-dwelling ants [Tapinoma sessile and
#>     Myrmica lobifrons]: 3.37), but on average hornets have greater than
#>     100 times the dry mass of these ants, and are easier to collect and
#>     process as a standardized food source. Additions of prey, either as
#>     large "pulses" or chronic "presses" are
#>     analogous to the enrichment and eutrophication that occur in aquatic
#>     "green" food webs in which phytoplankton abundance is
#>     boosted through addition of limiting nutrients. In "brown"
#>     food webs such as the Sarracenia microecosystem, detritus - not
#>     primary production - is at the base of the web, and our treatments
#>     boost this material as would happen through increases in arthropod
#>     prey capture78 or through nitrogen-enriched precipitation.
#>   </para>
#>     <para>
#>     Proteomic analysis. Proteomic profiles of microbial communities are
#>     determined after separating the microbial fraction from the pitcher
#>     fluid, prey, and other detritus. The microbial "pellet" is
#>     subjected to SDS-PAGE (sodium dodecyl sulfate polyacrylamide gel)
#>     electrophoresis; bands are cut out and digested in-gel with trypsin.
#>     Tryptic peptides are subjected to LC-MS/MS (liquid chromatography
#>     tandem mass spectrometry) for peptide and protein identification.
#>     Absolute abundance of peptides and proteins are quantified using
#>     AQUA (Absolute QUAntification) analysis109.
#>   </para>
#>   </section>
#>   <section>
#>     <title>Specific Experiments</title>
#>     <para>
#>     Experiment #1. Effects of nutrient enrichment on state changes and
#>     [O2] profiles. This experiment alters nutrient enrichment rates to
#>     characterize the [O2] profile and the transition to the anaerobic
#>     state. The experimental design is a one-way layout with 5 treatment
#>     groups: one control (no enrichment) and 4 enrichment levels (0.125,
#>     0.25, 0.5, 1.0 mg prey added ml-1 d-1). One plant is assigned to
#>     each treatment group, and the entire set is replicated 6 times over
#>     successive weeks. [O2] is monitored continuously for 4 days to
#>     characterize state changes and tipping points under different
#>     enrichment rates. This experiment tracks a continuous [O2] profile
#>     but does not include proteomic analysis. The purpose of Experiment
#>     #1 is to identify an enrichment rate E that generates a long
#>     pre-tipping period before transition time T to the anaerobic state.
#>     This enrichment rate will be used in Experiments #2 - #4.
#>   </para>
#>     <para>
#>     Experiment #2. Identification of early intervention time and
#>     characterization of aerobic and anaerobic proteomes. This experiment
#>     will use the single enrichment rate E determined from Experiment #1
#>     and impose different intervention times I at which nutrient
#>     enrichment will be terminated. Thus, this experiment will identify
#>     the latest time I* at which it is possible to intervene and stop the
#>     transition to the anaerobic state by halting enrichment. The [O2]
#>     profile will again be monitored continuously over 10 days to measure
#>     the state of the system. From Experiment #1, the transition time T
#>     to the anaerobic state with no intervention will be known. We will
#>     use one control group (no prey addition) and ten levels of
#>     intervention time (all with the same enrichment rate E) as a
#>     proportion of T (0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1.0).
#>     Six plants will be assigned randomly to each of the 11 treatments in
#>     a randomized one-way layout and [O2] profiles will be monitored
#>     continuously. In addition to the [O2] profiles, we will also
#>     characterize the protein profiles of aerobic and anaerobic pitchers
#>     in all 11 treatment groups at the end of the experiment.
#>   </para>
#>     <para>
#>     After the plants are harvested, we will create proteomic profiles of
#>     the predominantly bacterial portion (centrifuged pellet) of the
#>     pitcher fluid from each plant, as described in General Protocols.
#>     Thus, 66 separate pellet-fraction samples will be analyzed by
#>     SDS-PAGE. After examining the SDS-PAGE profiles, approximately ten
#>     proteins that show dynamic patterns consistent with state change and
#>     five that do not change will be cut from the gel, subjected to
#>     in-gel tryptic digestion and a portion of the tryptic peptides will
#>     be analyzed by LC-MS/MS. Using these data, we will choose three
#>     identified peptides from each protein for peptide synthesis such
#>     that each synthesized peptide contains stable isotope labels (AQUA
#>     peptides) to distinguish them by mass from the native peptides. We
#>     will then quantify all 45 of the native peptides from the original
#>     samples using a known amount of each AQUA peptide spiked into the
#>     tryptic digest. The AQUA analysis of proteins that do not show
#>     changes will be used for normalization between samples. These data
#>     will be used to independently identify the current state of the
#>     system and forecast the time-to-transition.
#>   </para>
#>     <para>
#>     We will use Sequest searches for initial identification of peptides;
#>     relevant scores including Xcorr and Delta Cn values will be given
#>     for each peptide. Other peptides will be identified by de novo
#>     sequencing using PepNovo; all PepNovo scores will likewise be given
#>     including any N- or C-terminal gaps. Mass error in ppm will be
#>     reported for each precursor ion. We will use standard multivariate
#>     analysis to search for distinctive proteomic profiles114 that
#>     characterize aerobic and anaerobic ecosystems, as well as ecosystems
#>     that developed with and without inputs of photosynthetic O2 and
#>     plant metabolites.
#>   </para>
#>     <para>
#>     Experiment #3. Identification of diagnostic proteins. Using
#>     Experiments #1 and 2, we will have identified an enrichment rate E
#>     with a long pre-tipping period and an intervention time I* before
#>     which mitigation and termination of enrichment will prevent
#>     eutrophication. Experiment #3 will characterize the mean and
#>     variance of the protein profile before and after I*. We are
#>     especially interested in identifying proteins that increase rapidly
#>     in abundance (or variance) well before the onset of flickering in
#>     [O2] and before the transition time T from the aerobic to the
#>     anaerobic state.
#>   </para>
#>     <para>
#>     A cohort of 100 plants all will be fed at rate E (determined from
#>     Experiment #1), with intervention time I* determined from Experiment
#>     #2, although no intervention will be used in this "press"
#>     experiment so that we can contrast proteins before and after the
#>     state change. At seven times before I* and three times after
#>     I<emphasis>, we will harvest 10 randomly chosen plants. At each
#>     prescribed harvest time, we will measure [O2] and collect samples
#>     from each plant for proteomic screening using both SDS-PAGE and AQUA
#>     analysis. This experiment will identify proteins that rise quickly
#>     in abundance during the pre- I</emphasis> period and can be used as
#>     early indicators of a future tipping point. Because different plants
#>     will be harvested at each time period, this is a one-way ANOVA
#>     design, with pre-and post- I* a priori contrasts. A randomization
#>     test will be used to determine whether variances in protein
#>     expression differ through time. During these analyses we will use
#>     the data from the AQUA peptides and from known amounts of protein
#>     standards, such as bovine serum albumin, to approximate the amount
#>     of protein in a given coomassie-stained SDS-PAGE gel band. The
#>     reason for doing this is to provide a fast "real-time"
#>     assay based just on expression in the SDS-PAGE. This rapid assay
#>     will be used in Experiment #4.
#>   </para>
#>     <para>
#>     Experiment # 4. Proof-of-application. This experiment will provide a
#>     benchmark test of our methods and their ability to correctly
#>     identify tipping points. A cohort of 100 plants will each be fitted
#>     with [O2] probes and started on the enrichment regime. Two times per
#>     day, we will collect 3 plants each, pool their contents, and conduct
#>     a rapid screen in the lab with SDS-PAGE for the diagnostic proteins
#>     that were identified in Experiment #3. We will use the protein
#>     expression in the gel to delineate an "early" and a
#>     "late" mitigation strategy. As soon as diagnostic proteins
#>     measured in the SDS-gels are at abundances that signal we are at
#>     0.5×I* - approximately one-half of the way to the latest
#>     intervention time - we will randomly select one third of the
#>     remaining plants for mitigation and termination of enrichment (the
#>     "early" mitigation strategy). We will continue to harvest
#>     plants from the remainder of the cohort and monitor proteins. As
#>     soon as diagnostic proteins signal we are at 0.75 times I*, we will
#>     randomly select one half of the remaining plants for mitigation and
#>     termination of enrichment (the "late" mitigation
#>     strategy). The remaining plants (approximately one sixth to one
#>     third of the original cohort) will continue to be enriched. We will
#>     monitor [O2] in all 3 groups (no-mitigation control, early
#>     mitigation, late mitigation) until all plants reach a new [O2]
#>     equilibrium. If the protein markers are successful, the proportion
#>     of food webs that remain aerobic will be significantly higher in the
#>     two mitigation treatments than in the no-mitigation control.
#>   </para>
#>   </section>
#> </description>