set_methods

set_methods

set_methods(methods_file, instrumentation = character(),
  software = NULL, sampling_file = NULL, sampling_coverage = NULL,
  sampling_citation = NULL, qualityControl_file = NULL)

Arguments

methods_file
Path to a file (markdown or .docx) containing a description of the methods used

instrumentation
optional, text describing instrumentation used in methods

software
optional, an EML software node describing software used in methods

sampling_file
optional, Path to a file (.md or .docx) describing sampling method

sampling_coverage
optional, coverage node for methods, e.g. set_coverage()

sampling_citation
optional, a citation element describing the sampling protocol

qualityControl_file
optional, path to a file (.md or .docx) describing quality control methods

Value

A methods object

Examples

f <- system.file("examples/hf205-methods.md", package = "EML")
set_methods(methods_file = f)

#> $sampling
#> NULL
#> 
#> $methodStep
#> $methodStep$instrumentation
#> character(0)
#> 
#> $methodStep$software
#> NULL
#> 
#> $methodStep$description
#> $methodStep$description$section
#> $methodStep$description$section[[1]]
#> [1] "<title>General Protocols</title>\n<para>\n    Field methods. All experiments will be carried out in the greenhouse\n    at Harvard Forest. We have developed an instrumentation system that\n    allows us to collect continuous dissolved [O2] measurements:\n    dedicated micro-probes (DO-166MT; Lazar Research Laboratories:\n    http://www.shelfscientific.com/) connected to multiplexers and data\n    loggers (AM16/32B multiplexer, CR-1000 datalogger and control system\n    [Campbell Scientific: http://www.cambellsci.com]). The initial\n    ecosystem composition in all experimental plants will be\n    standardized by seeding each pitcher with a 10-ml inoculum of liquid\n    collected from pitchers growing at Tom Swamp. In all experiments,\n    prey will be supplied to pitchers as standardized aliquots of dried\n    and finely ground bald-faced hornets (Dolichovespula maculata;\n    Hymenoptera: Vespidae), which we collect in quantity throughout New\n    England. Both hornets and ants (the latter are the dominant prey of\n    S. purpurea) are hymenoptera, and have nearly identical C:N ratios\n    (hornets: 3.97; common bog-dwelling ants [Tapinoma sessile and\n    Myrmica lobifrons]: 3.37), but on average hornets have greater than\n    100 times the dry mass of these ants, and are easier to collect and\n    process as a standardized food source. Additions of prey, either as\n    large <quote>pulses</quote> or chronic <quote>presses</quote> are\n    analogous to the enrichment and eutrophication that occur in aquatic\n    <quote>green</quote> food webs in which phytoplankton abundance is\n    boosted through addition of limiting nutrients. In\n    <quote>brown</quote> food webs such as the Sarracenia\n    microecosystem, detritus - not primary production - is at the base\n    of the web, and our treatments boost this material as would happen\n    through increases in arthropod prey capture78 or through\n    nitrogen-enriched precipitation.\n  </para>\n<para>\n    Proteomic analysis. Proteomic profiles of microbial communities are\n    determined after separating the microbial fraction from the pitcher\n    fluid, prey, and other detritus. The microbial <quote>pellet</quote>\n    is subjected to SDS-PAGE (sodium dodecyl sulfate polyacrylamide gel)\n    electrophoresis; bands are cut out and digested in-gel with trypsin.\n    Tryptic peptides are subjected to LC-MS/MS (liquid chromatography\n    tandem mass spectrometry) for peptide and protein identification.\n    Absolute abundance of peptides and proteins are quantified using\n    AQUA (Absolute QUAntification) analysis109.\n  </para>"
#> 
#> $methodStep$description$section[[2]]
#> [1] "<title>Specific Experiments</title>\n<para>\n    Experiment #1. Effects of nutrient enrichment on state changes and\n    [O2] profiles. This experiment alters nutrient enrichment rates to\n    characterize the [O2] profile and the transition to the anaerobic\n    state. The experimental design is a one-way layout with 5 treatment\n    groups: one control (no enrichment) and 4 enrichment levels (0.125,\n    0.25, 0.5, 1.0 mg prey added ml-1 d-1). One plant is assigned to\n    each treatment group, and the entire set is replicated 6 times over\n    successive weeks. [O2] is monitored continuously for 4 days to\n    characterize state changes and tipping points under different\n    enrichment rates. This experiment tracks a continuous [O2] profile\n    but does not include proteomic analysis. The purpose of Experiment\n    #1 is to identify an enrichment rate E that generates a long\n    pre-tipping period before transition time T to the anaerobic state.\n    This enrichment rate will be used in Experiments #2 - #4.\n  </para>\n<para>\n    Experiment #2. Identification of early intervention time and\n    characterization of aerobic and anaerobic proteomes. This experiment\n    will use the single enrichment rate E determined from Experiment #1\n    and impose different intervention times I at which nutrient\n    enrichment will be terminated. Thus, this experiment will identify\n    the latest time I* at which it is possible to intervene and stop the\n    transition to the anaerobic state by halting enrichment. The [O2]\n    profile will again be monitored continuously over 10 days to measure\n    the state of the system. From Experiment #1, the transition time T\n    to the anaerobic state with no intervention will be known. We will\n    use one control group (no prey addition) and ten levels of\n    intervention time (all with the same enrichment rate E) as a\n    proportion of T (0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1.0).\n    Six plants will be assigned randomly to each of the 11 treatments in\n    a randomized one-way layout and [O2] profiles will be monitored\n    continuously. In addition to the [O2] profiles, we will also\n    characterize the protein profiles of aerobic and anaerobic pitchers\n    in all 11 treatment groups at the end of the experiment.\n  </para>\n<para>\n    After the plants are harvested, we will create proteomic profiles of\n    the predominantly bacterial portion (centrifuged pellet) of the\n    pitcher fluid from each plant, as described in General Protocols.\n    Thus, 66 separate pellet-fraction samples will be analyzed by\n    SDS-PAGE. After examining the SDS-PAGE profiles, approximately ten\n    proteins that show dynamic patterns consistent with state change and\n    five that do not change will be cut from the gel, subjected to\n    in-gel tryptic digestion and a portion of the tryptic peptides will\n    be analyzed by LC-MS/MS. Using these data, we will choose three\n    identified peptides from each protein for peptide synthesis such\n    that each synthesized peptide contains stable isotope labels (AQUA\n    peptides) to distinguish them by mass from the native peptides. We\n    will then quantify all 45 of the native peptides from the original\n    samples using a known amount of each AQUA peptide spiked into the\n    tryptic digest. The AQUA analysis of proteins that do not show\n    changes will be used for normalization between samples. These data\n    will be used to independently identify the current state of the\n    system and forecast the time-to-transition.\n  </para>\n<para>\n    We will use Sequest searches for initial identification of peptides;\n    relevant scores including Xcorr and Delta Cn values will be given\n    for each peptide. Other peptides will be identified by de novo\n    sequencing using PepNovo; all PepNovo scores will likewise be given\n    including any N- or C-terminal gaps. Mass error in ppm will be\n    reported for each precursor ion. We will use standard multivariate\n    analysis to search for distinctive proteomic profiles114 that\n    characterize aerobic and anaerobic ecosystems, as well as ecosystems\n    that developed with and without inputs of photosynthetic O2 and\n    plant metabolites.\n  </para>\n<para>\n    Experiment #3. Identification of diagnostic proteins. Using\n    Experiments #1 and 2, we will have identified an enrichment rate E\n    with a long pre-tipping period and an intervention time I* before\n    which mitigation and termination of enrichment will prevent\n    eutrophication. Experiment #3 will characterize the mean and\n    variance of the protein profile before and after I*. We are\n    especially interested in identifying proteins that increase rapidly\n    in abundance (or variance) well before the onset of flickering in\n    [O2] and before the transition time T from the aerobic to the\n    anaerobic state.\n  </para>\n<para>\n    A cohort of 100 plants all will be fed at rate E (determined from\n    Experiment #1), with intervention time I* determined from Experiment\n    #2, although no intervention will be used in this\n    <quote>press</quote> experiment so that we can contrast proteins\n    before and after the state change. At seven times before I* and\n    three times after I<emphasis>, we will harvest 10 randomly chosen\n    plants. At each prescribed harvest time, we will measure [O2] and\n    collect samples from each plant for proteomic screening using both\n    SDS-PAGE and AQUA analysis. This experiment will identify proteins\n    that rise quickly in abundance during the pre- I</emphasis> period\n    and can be used as early indicators of a future tipping point.\n    Because different plants will be harvested at each time period, this\n    is a one-way ANOVA design, with pre-and post- I* a priori contrasts.\n    A randomization test will be used to determine whether variances in\n    protein expression differ through time. During these analyses we\n    will use the data from the AQUA peptides and from known amounts of\n    protein standards, such as bovine serum albumin, to approximate the\n    amount of protein in a given coomassie-stained SDS-PAGE gel band.\n    The reason for doing this is to provide a fast\n    <quote>real-time</quote> assay based just on expression in the\n    SDS-PAGE. This rapid assay will be used in Experiment #4.\n  </para>\n<para>\n    Experiment # 4. Proof-of-application. This experiment will provide a\n    benchmark test of our methods and their ability to correctly\n    identify tipping points. A cohort of 100 plants will each be fitted\n    with [O2] probes and started on the enrichment regime. Two times per\n    day, we will collect 3 plants each, pool their contents, and conduct\n    a rapid screen in the lab with SDS-PAGE for the diagnostic proteins\n    that were identified in Experiment #3. We will use the protein\n    expression in the gel to delineate an <quote>early</quote> and a\n    <quote>late</quote> mitigation strategy. As soon as diagnostic\n    proteins measured in the SDS-gels are at abundances that signal we\n    are at 0.5×I* - approximately one-half of the way to the latest\n    intervention time - we will randomly select one third of the\n    remaining plants for mitigation and termination of enrichment (the\n    <quote>early</quote> mitigation strategy). We will continue to\n    harvest plants from the remainder of the cohort and monitor\n    proteins. As soon as diagnostic proteins signal we are at 0.75 times\n    I*, we will randomly select one half of the remaining plants for\n    mitigation and termination of enrichment (the <quote>late</quote>\n    mitigation strategy). The remaining plants (approximately one sixth\n    to one third of the original cohort) will continue to be enriched.\n    We will monitor [O2] in all 3 groups (no-mitigation control, early\n    mitigation, late mitigation) until all plants reach a new [O2]\n    equilibrium. If the protein markers are successful, the proportion\n    of food webs that remain aerobic will be significantly higher in the\n    two mitigation treatments than in the no-mitigation control.\n  </para>"
#> 
#> 
#> $methodStep$description$para
#> list()
#> 
#> 
#> 

## Can also import from methods written in a .docx MS Word file.
f <- system.file("examples/hf205-methods.docx", package = "EML")
set_methods(methods_file = f)

#> $sampling
#> NULL
#> 
#> $methodStep
#> $methodStep$instrumentation
#> character(0)
#> 
#> $methodStep$software
#> NULL
#> 
#> $methodStep$description
#> $methodStep$description$section
#> $methodStep$description$section[[1]]
#> [1] "<title>General Protocols</title>\n<para>\n    Field methods. All experiments will be carried out in the greenhouse\n    at Harvard Forest. We have developed an instrumentation system that\n    allows us to collect continuous dissolved [O2] measurements:\n    dedicated micro-probes (DO-166MT; Lazar Research Laboratories:\n    http://www.shelfscientific.com/) connected to multiplexers and data\n    loggers (AM16/32B multiplexer, CR-1000 datalogger and control system\n    [Campbell Scientific: http://www.cambellsci.com]). The initial\n    ecosystem composition in all experimental plants will be\n    standardized by seeding each pitcher with a 10-ml inoculum of liquid\n    collected from pitchers growing at Tom Swamp. In all experiments,\n    prey will be supplied to pitchers as standardized aliquots of dried\n    and finely ground bald-faced hornets (Dolichovespula maculata;\n    Hymenoptera: Vespidae), which we collect in quantity throughout New\n    England. Both hornets and ants (the latter are the dominant prey of\n    S. purpurea) are hymenoptera, and have nearly identical C:N ratios\n    (hornets: 3.97; common bog-dwelling ants [Tapinoma sessile and\n    Myrmica lobifrons]: 3.37), but on average hornets have greater than\n    100 times the dry mass of these ants, and are easier to collect and\n    process as a standardized food source. Additions of prey, either as\n    large \"pulses\" or chronic \"presses\" are\n    analogous to the enrichment and eutrophication that occur in aquatic\n    \"green\" food webs in which phytoplankton abundance is\n    boosted through addition of limiting nutrients. In \"brown\"\n    food webs such as the Sarracenia microecosystem, detritus - not\n    primary production - is at the base of the web, and our treatments\n    boost this material as would happen through increases in arthropod\n    prey capture78 or through nitrogen-enriched precipitation.\n  </para>\n<para>\n    Proteomic analysis. Proteomic profiles of microbial communities are\n    determined after separating the microbial fraction from the pitcher\n    fluid, prey, and other detritus. The microbial \"pellet\" is\n    subjected to SDS-PAGE (sodium dodecyl sulfate polyacrylamide gel)\n    electrophoresis; bands are cut out and digested in-gel with trypsin.\n    Tryptic peptides are subjected to LC-MS/MS (liquid chromatography\n    tandem mass spectrometry) for peptide and protein identification.\n    Absolute abundance of peptides and proteins are quantified using\n    AQUA (Absolute QUAntification) analysis109.\n  </para>"
#> 
#> $methodStep$description$section[[2]]
#> [1] "<title>Specific Experiments</title>\n<para>\n    Experiment #1. Effects of nutrient enrichment on state changes and\n    [O2] profiles. This experiment alters nutrient enrichment rates to\n    characterize the [O2] profile and the transition to the anaerobic\n    state. The experimental design is a one-way layout with 5 treatment\n    groups: one control (no enrichment) and 4 enrichment levels (0.125,\n    0.25, 0.5, 1.0 mg prey added ml-1 d-1). One plant is assigned to\n    each treatment group, and the entire set is replicated 6 times over\n    successive weeks. [O2] is monitored continuously for 4 days to\n    characterize state changes and tipping points under different\n    enrichment rates. This experiment tracks a continuous [O2] profile\n    but does not include proteomic analysis. The purpose of Experiment\n    #1 is to identify an enrichment rate E that generates a long\n    pre-tipping period before transition time T to the anaerobic state.\n    This enrichment rate will be used in Experiments #2 - #4.\n  </para>\n<para>\n    Experiment #2. Identification of early intervention time and\n    characterization of aerobic and anaerobic proteomes. This experiment\n    will use the single enrichment rate E determined from Experiment #1\n    and impose different intervention times I at which nutrient\n    enrichment will be terminated. Thus, this experiment will identify\n    the latest time I* at which it is possible to intervene and stop the\n    transition to the anaerobic state by halting enrichment. The [O2]\n    profile will again be monitored continuously over 10 days to measure\n    the state of the system. From Experiment #1, the transition time T\n    to the anaerobic state with no intervention will be known. We will\n    use one control group (no prey addition) and ten levels of\n    intervention time (all with the same enrichment rate E) as a\n    proportion of T (0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1.0).\n    Six plants will be assigned randomly to each of the 11 treatments in\n    a randomized one-way layout and [O2] profiles will be monitored\n    continuously. In addition to the [O2] profiles, we will also\n    characterize the protein profiles of aerobic and anaerobic pitchers\n    in all 11 treatment groups at the end of the experiment.\n  </para>\n<para>\n    After the plants are harvested, we will create proteomic profiles of\n    the predominantly bacterial portion (centrifuged pellet) of the\n    pitcher fluid from each plant, as described in General Protocols.\n    Thus, 66 separate pellet-fraction samples will be analyzed by\n    SDS-PAGE. After examining the SDS-PAGE profiles, approximately ten\n    proteins that show dynamic patterns consistent with state change and\n    five that do not change will be cut from the gel, subjected to\n    in-gel tryptic digestion and a portion of the tryptic peptides will\n    be analyzed by LC-MS/MS. Using these data, we will choose three\n    identified peptides from each protein for peptide synthesis such\n    that each synthesized peptide contains stable isotope labels (AQUA\n    peptides) to distinguish them by mass from the native peptides. We\n    will then quantify all 45 of the native peptides from the original\n    samples using a known amount of each AQUA peptide spiked into the\n    tryptic digest. The AQUA analysis of proteins that do not show\n    changes will be used for normalization between samples. These data\n    will be used to independently identify the current state of the\n    system and forecast the time-to-transition.\n  </para>\n<para>\n    We will use Sequest searches for initial identification of peptides;\n    relevant scores including Xcorr and Delta Cn values will be given\n    for each peptide. Other peptides will be identified by de novo\n    sequencing using PepNovo; all PepNovo scores will likewise be given\n    including any N- or C-terminal gaps. Mass error in ppm will be\n    reported for each precursor ion. We will use standard multivariate\n    analysis to search for distinctive proteomic profiles114 that\n    characterize aerobic and anaerobic ecosystems, as well as ecosystems\n    that developed with and without inputs of photosynthetic O2 and\n    plant metabolites.\n  </para>\n<para>\n    Experiment #3. Identification of diagnostic proteins. Using\n    Experiments #1 and 2, we will have identified an enrichment rate E\n    with a long pre-tipping period and an intervention time I* before\n    which mitigation and termination of enrichment will prevent\n    eutrophication. Experiment #3 will characterize the mean and\n    variance of the protein profile before and after I*. We are\n    especially interested in identifying proteins that increase rapidly\n    in abundance (or variance) well before the onset of flickering in\n    [O2] and before the transition time T from the aerobic to the\n    anaerobic state.\n  </para>\n<para>\n    A cohort of 100 plants all will be fed at rate E (determined from\n    Experiment #1), with intervention time I* determined from Experiment\n    #2, although no intervention will be used in this \"press\"\n    experiment so that we can contrast proteins before and after the\n    state change. At seven times before I* and three times after\n    I<emphasis>, we will harvest 10 randomly chosen plants. At each\n    prescribed harvest time, we will measure [O2] and collect samples\n    from each plant for proteomic screening using both SDS-PAGE and AQUA\n    analysis. This experiment will identify proteins that rise quickly\n    in abundance during the pre- I</emphasis> period and can be used as\n    early indicators of a future tipping point. Because different plants\n    will be harvested at each time period, this is a one-way ANOVA\n    design, with pre-and post- I* a priori contrasts. A randomization\n    test will be used to determine whether variances in protein\n    expression differ through time. During these analyses we will use\n    the data from the AQUA peptides and from known amounts of protein\n    standards, such as bovine serum albumin, to approximate the amount\n    of protein in a given coomassie-stained SDS-PAGE gel band. The\n    reason for doing this is to provide a fast \"real-time\"\n    assay based just on expression in the SDS-PAGE. This rapid assay\n    will be used in Experiment #4.\n  </para>\n<para>\n    Experiment # 4. Proof-of-application. This experiment will provide a\n    benchmark test of our methods and their ability to correctly\n    identify tipping points. A cohort of 100 plants will each be fitted\n    with [O2] probes and started on the enrichment regime. Two times per\n    day, we will collect 3 plants each, pool their contents, and conduct\n    a rapid screen in the lab with SDS-PAGE for the diagnostic proteins\n    that were identified in Experiment #3. We will use the protein\n    expression in the gel to delineate an \"early\" and a\n    \"late\" mitigation strategy. As soon as diagnostic proteins\n    measured in the SDS-gels are at abundances that signal we are at\n    0.5×I* - approximately one-half of the way to the latest\n    intervention time - we will randomly select one third of the\n    remaining plants for mitigation and termination of enrichment (the\n    \"early\" mitigation strategy). We will continue to harvest\n    plants from the remainder of the cohort and monitor proteins. As\n    soon as diagnostic proteins signal we are at 0.75 times I*, we will\n    randomly select one half of the remaining plants for mitigation and\n    termination of enrichment (the \"late\" mitigation\n    strategy). The remaining plants (approximately one sixth to one\n    third of the original cohort) will continue to be enriched. We will\n    monitor [O2] in all 3 groups (no-mitigation control, early\n    mitigation, late mitigation) until all plants reach a new [O2]\n    equilibrium. If the protein markers are successful, the proportion\n    of food webs that remain aerobic will be significantly higher in the\n    two mitigation treatments than in the no-mitigation control.\n  </para>"
#> 
#> 
#> $methodStep$description$para
#> list()
#> 
#> 
#>