set_taxonomicCoverage

set_taxonomicCoverage

set_taxonomicCoverage(sci_names)

Arguments

sci_names

string (space seperated) or list or data frame of scientific names for species covered.

Value

a taxonomicCoverage object for EML

Details

Turn a data.frame or a list of scientific names into a taxonomicCoverage block sci_names can be a space-separated character string or a data frame with column names as rank name or a list of user-defined taxonomicClassification

Note

If "sci_names" is a data frame, column names of the data frame are rank names. For user-defined "sci_names", users must make sure that the order of rank names they specify is from high to low. Ex. "Kingdom","Phylum","Class","Order","Family","Genus","Species","Common"

Examples

taxon_coverage <-
 set_taxonomicCoverage(list(KINGDOM="Plantae",
                            PHYLUM="Phaeophyta",
                            CLASS="Phaeophyceae",
                            ORDER="Laminariales",
                            FAMILY="Lessoniaceae",
                            GENUS="Macrocystis",
                            genusSpecies="Macrocystis pyrifera",
                            commonName="MAPY"))

df <- data.frame(KINGDOM="Plantae",
                 PHYLUM="Phaeophyta",
                 CLASS="Phaeophyceae",
                 ORDER="Laminariales",
                 FAMILY="Lessoniaceae",
                 GENUS="Macrocystis",
                 genusSpecies="Macrocystis pyrifera",
                 commonName="MAPY")
taxon_coverage <- set_taxonomicCoverage(df)